feat: support hot-reload for ui
If starting app with `nrtk_explorer --hot-reload`,
then a button appears in the upper right to reload
the ui modules and re-run Engine.ui method.
